requirements.txt
# n.process API - Python Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Google Cloud Platform
google-auth==2.35.0
google-cloud-firestore==2.19.0
google-cloud-aiplatform==1.71.1

# Vertex AI
vertexai==1.71.1

# Utilities
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2

# Backup & Restore
google-cloud-storage==2.14.0

# Security
bcrypt==4.2.0

# Production
gunicorn==23.0.0

# Rate Limiting (optional - falls back to in-memory if not available)
redis==5.0.1

# Observability (optional)
google-cloud-logging==3.11.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-cloud-trace==0.10b1
opentelemetry-instrumentation-fastapi==0.45b0
google-cloud-monitoring==2.19.0

# Secret Management
google-cloud-secret-manager==2.20.0

# Firebase Admin SDK for token verification
firebase-admin==6.6.0

# Ingestion Engine
langchain==0.3.0
langchain-text-splitters==0.3.0
beautifulsoup4==4.12.3
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
